IKASAN-2644 changing the scheduler agent so that it has a fixed number of file processing flows in order to allow for better scaling of the applications ability to detect file arrivals.

+++ ikasaneip/ootb/module/scheduler-agent/jar/src/test/java/org/ikasan/ootb/scheduler/agent/module/component/broker/MoveFileBrokerTest.java
@@ -2,79 +2,74 @@ package org.ikasan.ootb.scheduler.agent.module.component.broker;
 
 import org.apache.commons.io.FileUtils;
 import org.ikasan.component.endpoint.filesystem.messageprovider.CorrelatedFileList;
-import org.ikasan.ootb.scheduler.agent.module.component.broker.configuration.MoveFileBrokerConfiguration;
-import org.ikasan.spec.component.endpoint.EndpointException;
-import org.ikasan.spec.scheduled.dryrun.DryRunModeService;
+import org.ikasan.ootb.scheduler.agent.module.model.FileWatcherJobEvent;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-
 @RunWith(MockitoJUnitRunner.class)
 public class MoveFileBrokerTest {
 
-    @Mock
-    DryRunModeService dryRunModeService;
-
-    @Test(expected = IllegalArgumentException.class)
-    public void test_exception_constructor_null_dry_run_service() {
-        new MoveFileBroker(null);
-    }
-
     @Before
-    public void createArchiveDirectory() {
+    public void createArchiveDirectory() throws IOException {
         File archiveDirectory = new File("src/test/resources/data/archive");
         if (!archiveDirectory.exists()) {
             archiveDirectory.mkdirs();
         }
+
+        if(new File("src/test/resources/data/archive/test.txt").exists()) {
+            FileUtils.moveFileToDirectory(new File("src/test/resources/data/archive/test.txt")
+                , new File("src/test/resources/data"), true);
+        }
+        FileUtils.cleanDirectory(new File("src/test/resources/data/archive"));
     }
 
     @Test
     public void test_move_file_success() throws IOException {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
+       List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        List<File> files = List.of(new File("src/test/resources/data/test.txt"));
-
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory("src/test/resources/data/archive");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory("src/test/resources/data/archive");
+        event.setJobName("jobName");
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
         Assert.assertEquals("test.txt", archiveFiles[0]);
-
-        FileUtils.moveFileToDirectory(new File("src/test/resources/data/archive/test.txt")
-            , new File("src/test/resources/data"), true);
     }
 
     @Test
     public void test_same_file_twice_move_file_success() throws IOException {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory("src/test/resources/data/archive");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory("src/test/resources/data/archive");
+        event.setJobName("jobName");
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
@@ -83,35 +78,31 @@ public class MoveFileBrokerTest {
         FileUtils.copyFileToDirectory(new File("src/test/resources/data/archive/test.txt")
             , new File("src/test/resources/data"), true);
 
-        broker.invoke(correlatedFileList);
+        broker.invoke(event);
 
         archiveFiles = new File("src/test/resources/data/archive").list();
 
         Assert.assertEquals(2, archiveFiles.length);
-
-        FileUtils.moveFileToDirectory(new File("src/test/resources/data/archive/test.txt")
-            , new File("src/test/resources/data"), true);
-        FileUtils.cleanDirectory(new File("src/test/resources/data/archive"));
-
-        archiveFiles = new File("src/test/resources/data/archive").list();
-
-        Assert.assertEquals(0, archiveFiles.length);
     }
 
     @Test
     public void test_move_file_dry_run_success() {
-        when(dryRunModeService.getDryRunMode()).thenReturn(true);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory("src/test/resources/data/archive");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory("src/test/resources/data/archive");
+        event.setJobName("jobName");
+        event.setDryRun(true);
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
@@ -120,19 +111,22 @@ public class MoveFileBrokerTest {
 
     @Test
     public void test_move_file_job_dry_run_success() {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
-        when(dryRunModeService.isJobDryRun(anyString())).thenReturn(true);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory("src/test/resources/data/archive");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory("src/test/resources/data/archive");
+        event.setJobName("jobName");
+        event.setDryRun(true);
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
@@ -141,18 +135,21 @@ public class MoveFileBrokerTest {
 
     @Test
     public void test_move_file_dry_run_exception_bad_move_directory_same_as_src_directory() throws IOException {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory("src/test/resources/data/");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory("src/test/resources/data/");
+        event.setJobName("jobName");
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
@@ -161,18 +158,21 @@ public class MoveFileBrokerTest {
 
     @Test
     public void test_move_file_dry_run_exception_bad_move_directory_same_as_src_directory_due_to_dot() throws IOException {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setMoveDirectory(".");
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setMoveDirectory(".");
+        event.setJobName("jobName");
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
 
@@ -181,17 +181,20 @@ public class MoveFileBrokerTest {
 
     @Test
     public void test_move_no_target_dir_success() throws IOException {
-        when(dryRunModeService.getDryRunMode()).thenReturn(false);
-
         List<File> files = List.of(new File("src/test/resources/data/test.txt"));
 
-        MoveFileBrokerConfiguration configuration = new MoveFileBrokerConfiguration();
-        configuration.setJobName("jobName");
-        MoveFileBroker broker = new MoveFileBroker(this.dryRunModeService);
-        broker.setConfiguration(configuration);
+        MoveFileBroker broker = new MoveFileBroker();
 
         CorrelatedFileList correlatedFileList = new CorrelatedFileList(files, "correlationIdentifier");
-        broker.invoke(correlatedFileList);
+
+        FileWatcherJobEvent event = new FileWatcherJobEvent();
+        event.setCorrelationIdentifier("correlationIdentifier");
+        event.setCorrelatedFileList(correlatedFileList);
+        event.setJobName("jobName");
+        event.setMinFileAgeSeconds(30);
+        event.setJobName("jobName");
+
+        broker.invoke(event);
 
         String[] archiveFiles = new File("src/test/resources/data/archive").list();
